% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geno_mat.R
\docType{data}
\name{geno_mat}
\alias{geno_mat}
\title{Model Gene Compendium Data Set}
\format{
A numeric matrix with 1345 rows and 2271 columns:
}
\source{
{Generated by Zachary James Winn for the CSU breeding program via historical in-house GBS data}
}
\usage{
geno_mat
}
\description{
A numeric matrix which contains molecular marker information on 1345 unique genotypes for 2271 SNP markers located on wheat chromosome 3B. This data set corresponds to the information found in the "gene_comp" and "marker_info" data sets.
}
\examples{
data("geno_mat") #lazy loads the dataset for use in the package
}
\keyword{datasets}
