% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locus_pred.R
\name{geno_pred}
\alias{geno_pred}
\title{Haplotype Prediction: Using Trained Models to Make Predictions}
\usage{
geno_pred(geno_train_results, geno_mat, genotypes_to_predict)
}
\arguments{
\item{geno_mat}{This is a genotypic matrix with genotypes of individuals you have genotyped and characterized for the locus/QTL/gene of interest and individuals that have only been genotyped with a genome wide marker panel. It is important to note that the markers in the genome wide panel \emph{must} be shared between training population and the population you wish to forward predict. In the case of genotyping-by-sequencing markers, the training and test populations should be discovered and produced together into a genotype file. All marker platforms, however, are compatable, as long as the training and forward prediction population share the same markers genome-wide.}

\item{genotypes_to_predict}{This is a character vector of genotypes in the geno_mat which the user wishes to predict. If this object contains names in the training population, they will be omitted in the results to avoid bias.}

\item{locus_train_results}{This object is a the results object from the "locus_train" function.}
}
\value{
a data frame with two columns: genotype names and predictions.
}
\description{
Haplotype Prediction: Using Trained Models to Make Predictions
}
