% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marker_info.R
\docType{data}
\name{marker_info}
\alias{marker_info}
\title{Model Gene Compendium Data Set}
\format{
A data frame with 2271 rows and 3 columns:

\describe{

\item{Marker}{The designation of the markers which are found in the genotypic matrix}
\item{Chromosome}{The chromosome where each marker resides}
\item{BP_Position}{The position of each marker in basepairs}

}
}
\source{
{Generated by Zachary James Winn for the CSU breeding program via historical in-house GBS data}
}
\usage{
marker_info
}
\description{
A data frame which contains marker information of GBS markers found on wheat chromosome 3B. This data pairs with the markers found in "geno_mat" data file associated with the HaploCatcher package.
}
\examples{
data("marker_info") #lazy loads the dataset for use in the package
}
\keyword{datasets}
