% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_locus_perm_cv.R
\name{plot_locus_perm_cv}
\alias{plot_locus_perm_cv}
\title{Visualize Permutation CV Results}
\usage{
plot_locus_perm_cv(results, individual_images = FALSE)
}
\arguments{
\item{results}{An object of class "list" which is derived from the "locus_perm_cv" function.}

\item{individual_images}{A logical argument that defines if the user wants both the composite image and the full image. Default setting is FALSE.}
}
\value{
Prints a ggplot2 image
}
\description{
This function takes a list of list object from the "locus_perm_cv" function and creates a summary graphic of the accuracy, kappa, sensitivity, and specificity of the models ran. If heterozygous individuals were left in the cross validation scheme, by-class sensitivity and specificity will be displayed; otherwise, displayed parameters will be of the overall model.
}
\examples{
#read in the genotypic data matrix
data("geno_mat")

#read in the marker information
data("marker_info")

#read in the gene compendium file
data("gene_comp")

#run permutational analysis
fit<-locus_perm_cv(n_perms = 3, #the number of permutations
                  geno_mat=geno_mat, #the genotypic matrix
                  gene_file=gene_comp, #the gene compendium file
                  gene_name="sst1_solid_stem", #the name of the gene
                  marker_info=marker_info, #the marker information file
                  chromosome="3B", #name of the chromosome
                  ncor_markers=50, #number of markers to retain
                  percent_testing=0.2, #percentage of genotypes in the validation set
                  percent_training=0.8, #percentage of genotypes in the training set
                  include_hets=TRUE, #includes hets in the model
                  include_models=TRUE, #includes models in results object
                  verbose = TRUE) #includes text/plots

#plot results
plot_locus_perm_cv(fit,
                  individual_images=TRUE)

}
